jsutil: start splitting into separate source files, add LoadScript and LoadStylesheet
